+++ src/test/scala/io/archivesunleashed/app/ExtractPopularImagesTestRDD.scala
@@ -23,7 +23,7 @@ import org.scalatest.junit.JUnitRunner
 import org.scalatest.{BeforeAndAfter, FunSuite}
 
 @RunWith(classOf[JUnitRunner])
-class ExtractPopularImagesTest extends FunSuite with BeforeAndAfter {
+class ExtractPopularImagesTestRDD extends FunSuite with BeforeAndAfter {
   private val arcPath = Resources.getResource("arc/example.arc.gz").getPath
   private var sc: SparkContext = _
   private val master = "local[4]"
@@ -42,8 +42,8 @@ class ExtractPopularImagesTest extends FunSuite with BeforeAndAfter {
   test("extracts popular images") {
     val highTest = 507
     val examplerdd = RecordLoader.loadArchives(arcPath, sc)
-    val imagesLowLimit = ExtractPopularImages(examplerdd, 3, sc)
-    val imagesHighLimit = ExtractPopularImages(examplerdd, highTest, sc)
+    val imagesLowLimit = ExtractPopularImagesRDD(examplerdd, 3, sc)
+    val imagesHighLimit = ExtractPopularImagesRDD(examplerdd, highTest, sc)
     val response = Array("1\thttp://www.archive.org/images/books-small.jpg",
       "1\thttp://i.creativecommons.org/l/by-sa/3.0/88x31.png",
       "1\thttp://www.archive.org/images/blendbar.jpg")
